Dropped the refactoring according to offline discussions.
